[Feature] #45 : Selected Annotation UI
