+++ DaeguGoodPriceShop/DaeguGoodPriceShop/View/AnnotationView/KoreanFoodAnnotationView.swift
@@ -1,5 +1,5 @@
 //
-//  WesternFoodAnnotationView.swift
+//  KoreanFoodAnnotationView.swift
 //  DaeguGoodPriceShop
 //
 //  Created by Shin Jae Ung on 2022/07/07.
@@ -7,7 +7,7 @@
 
 import MapKit
 
-final class WesternFoodAnnotationView: ShopAnnotationView {
+final class KoreanFoodAnnotationView: ShopAnnotationView {
     override init(annotation: MKAnnotation?, reuseIdentifier: String?) {
         super.init(annotation: annotation, reuseIdentifier: reuseIdentifier)
         self.clusteringIdentifier = "CateringStore"
@@ -25,11 +25,11 @@ final class WesternFoodAnnotationView: ShopAnnotationView {
     
     override func prepareForDisplay() {
         super.prepareForDisplay()
-        self.image = resizedImage(image: UIImage(named: "westernFood"), width: 40, height: 40)
+        self.image = resizedImage(image: UIImage(named: "koreanFood"), width: 40, height: 40)
     }
-    
+
     override func selected() {
         super.selected()
-        self.image = resizedImage(image: UIImage(named: "westernFoodSelected"), width: 40, height: 49.92)
+        self.image = resizedImage(image: UIImage(named: "koreanFoodSelected"), width: 40, height: 49.92)
     }
 }
