+++ DaeguGoodPriceShop/DaeguGoodPriceShop/View/AnnotationView/WesternFoodAnnotationView.swift
@@ -1,5 +1,5 @@
 //
-//  KoreanFoodAnnotationView.swift
+//  WesternFoodAnnotationView.swift
 //  DaeguGoodPriceShop
 //
 //  Created by Shin Jae Ung on 2022/07/07.
@@ -7,7 +7,7 @@
 
 import MapKit
 
-final class KoreanFoodAnnotationView: ShopAnnotationView {
+final class WesternFoodAnnotationView: ShopAnnotationView {
     override init(annotation: MKAnnotation?, reuseIdentifier: String?) {
         super.init(annotation: annotation, reuseIdentifier: reuseIdentifier)
         self.clusteringIdentifier = "CateringStore"
@@ -25,11 +25,11 @@ final class KoreanFoodAnnotationView: ShopAnnotationView {
     
     override func prepareForDisplay() {
         super.prepareForDisplay()
-        self.image = resizedImage(image: UIImage(named: "koreanFood"), width: 40, height: 40)
+        self.image = resizedImage(image: UIImage(named: "westernFood"), width: 40, height: 40)
     }
-    
+
     override func selected() {
         super.selected()
-        self.image = resizedImage(image: UIImage(named: "koreanFoodSelected"), width: 40, height: 49.92)
+        self.image = resizedImage(image: UIImage(named: "westernFoodSelected"), width: 40, height: 49.92)
     }
 }
